Refactor SettingsFragment
